refactor: Explode the configurator-state giant file
